AsyncEnumerableSpec
