+++ src/core/Akka.Streams.Tests/Dsl/AsyncEnumerableSpec.cs
@@ -18,6 +18,11 @@ using FluentAssertions;
 using Nito.AsyncEx.Synchronous;
 using Xunit;
 using Xunit.Abstractions;
+using System.Collections.Generic;
+using Akka.Actor;
+using Akka.Streams.Actors;
+using Akka.Streams.Tests.Actor;
+using Reactive.Streams;
 
 namespace Akka.Streams.Tests.Dsl
 {
@@ -146,7 +151,63 @@ namespace Akka.Streams.Tests.Dsl
             
             await Assert.ThrowsAsync<IllegalStateException>(ShouldThrow);
         }
+        [Fact]
+        public async Task AsyncEnumerableSource_must_allow_multiple_IAsyncEnumerable()
+        {
+            var materializer = Sys.Materializer();
+            await this.AssertAllStagesStoppedAsync(async () =>
+            {
+                // creating actor with default supervision, because stream supervisor default strategy is to 
+                var actorRef = Sys.ActorOf(ManualSubscriber.Props(TestActor));
+                Source.AsyncEnumerableAsRun(RangeAsync(1, 7))
+                    .RunWith(Sink.FromSubscriber(new ActorSubscriberImpl<int>(actorRef)), materializer);
+                actorRef.Tell("ready");
+                (await ExpectMsgAsync<OnNext>()).Element.Should().Be(1);
+                (await ExpectMsgAsync<OnNext>()).Element.Should().Be(2);
+                await ExpectNoMsgAsync(200);
+                actorRef.Tell("boom");
+                actorRef.Tell("ready");
+                actorRef.Tell("ready");
+                actorRef.Tell("boom");
+                await foreach (var n in RangeAsync(3, 4))
+                {
+                    (await ExpectMsgAsync<OnNext>()).Element.Should().Be(n);
+                }
+                await ExpectNoMsgAsync(200);
+                actorRef.Tell("ready");
+                (await ExpectMsgAsync<OnNext>()).Element.Should().Be(7);
+                await ExpectMsgAsync<OnComplete>();
+            }, materializer);
+        }
         
+        [Fact]
+        public void AsyncEnumerableSource_Must_Complete_Immediately_With_No_elements_When_An_Empty_IAsyncEnumerable_Is_Passed_In()
+        {
+            var p = Source.AsyncEnumerableAsRun(AsyncEnumerable.Empty<int>()).RunWith(Sink.AsPublisher<int>(false), Sys.Materializer());
+            var c = this.CreateManualSubscriberProbe<int>();
+            p.Subscribe(c);
+            c.ExpectSubscriptionAndComplete();
+        }
+
+        [Fact]
+        public void AsyncEnumerableSource_Select()
+        {
+            var materializer = Sys.Materializer();
+            var p = Source.AsyncEnumerableAsRun(RangeAsync(1, 100))
+                    .Select(message =>
+                    {
+                        return message;
+                    });
+            ///////////
+        }
+        static async IAsyncEnumerable<int> RangeAsync(int start, int count)
+        {
+            for (int i = 0; i < count; i++)
+            {
+                await Task.Delay(i);
+                yield return start + i;
+            }
+        }
         
     }
